cache: remove no longer needed get_missing_housenumbers_html()
Because it used write_missing_housenumbers() internally, which already
uses the JSON cache. This way the locale-dependent HTML cache can be
removed. Includes:

- fix wsgi::tests::test_missing_housenumbers_compat
- remove now unused get_housenumbers_htmlcache_path() and
  get_housenumbers_htmlcache_read_stream() in area_files
