+++ src/cache/tests.rs
@@ -17,214 +17,6 @@ use std::collections::HashMap;
 use std::rc::Rc;
 use std::sync::Arc;
 
-/// Tests is_missing_housenumbers_html_cached().
-#[test]
-fn test_is_missing_housenumbers_html_cached() {
-    let mut ctx = context::tests::make_test_context().unwrap();
-    let mut file_system = context::tests::TestFileSystem::new();
-    let percent_value = context::tests::TestFileSystem::make_file();
-    let html_cache_value = context::tests::TestFileSystem::make_file();
-    let json_cache_value = context::tests::TestFileSystem::make_file();
-    let files = context::tests::TestFileSystem::make_files(
-        &ctx,
-        &[
-            ("workdir/gazdagret.percent", &percent_value),
-            ("workdir/gazdagret.htmlcache.en", &html_cache_value),
-            ("workdir/gazdagret.cache.json", &json_cache_value),
-        ],
-    );
-    file_system.set_files(&files);
-    let mut mtimes: HashMap<String, Rc<RefCell<f64>>> = HashMap::new();
-    mtimes.insert(
-        ctx.get_abspath("workdir/gazdagret.htmlcache.en"),
-        Rc::new(RefCell::new(0_f64)),
-    );
-    mtimes.insert(
-        ctx.get_abspath("workdir/gazdagret.cache.json"),
-        Rc::new(RefCell::new(0_f64)),
-    );
-    file_system.set_mtimes(&mtimes);
-    let file_system_arc: Arc<dyn context::FileSystem> = Arc::new(file_system);
-    ctx.set_file_system(&file_system_arc);
-    let mut relations = areas::Relations::new(&ctx).unwrap();
-    let mut relation = relations.get_relation("gazdagret").unwrap();
-    get_missing_housenumbers_html(&ctx, &mut relation).unwrap();
-
-    assert_eq!(
-        is_missing_housenumbers_html_cached(&ctx, &mut relation).unwrap(),
-        true
-    );
-}
-
-/// Tests is_missing_housenumbers_html_cached(): the case when there is no cache.
-#[test]
-fn test_is_missing_housenumbers_html_cached_no_cache() {
-    let mut ctx = context::tests::make_test_context().unwrap();
-    let mut file_system = context::tests::TestFileSystem::new();
-    let relation_percent = context::tests::TestFileSystem::make_file();
-    let relation_htmlcache = context::tests::TestFileSystem::make_file();
-    let files = context::tests::TestFileSystem::make_files(
-        &ctx,
-        &[
-            ("workdir/gazdagret.percent", &relation_percent),
-            ("workdir/gazdagret.htmlcache.en", &relation_htmlcache),
-        ],
-    );
-    file_system.set_files(&files);
-    let mut mtimes: HashMap<String, Rc<RefCell<f64>>> = HashMap::new();
-    mtimes.insert(
-        ctx.get_abspath("workdir/gazdagret.htmlcache.en"),
-        Rc::new(RefCell::new(0_f64)),
-    );
-    file_system.set_mtimes(&mtimes);
-    let file_system_arc: Arc<dyn context::FileSystem> = Arc::new(file_system);
-    ctx.set_file_system(&file_system_arc);
-    let mut relations = areas::Relations::new(&ctx).unwrap();
-    let relation = relations.get_relation("gazdagret").unwrap();
-
-    let is_cached = is_missing_housenumbers_html_cached(&ctx, &relation).unwrap();
-
-    assert_eq!(is_cached, false);
-}
-
-/// Tests is_missing_housenumbers_html_cached(): the case when osm_housenumbers is new, so the cache entry is old.
-#[test]
-fn test_is_missing_housenumbers_html_cached_osm_housenumbers_new() {
-    let mut ctx = context::tests::make_test_context().unwrap();
-    let mut relations = areas::Relations::new(&ctx).unwrap();
-    let relation = relations.get_relation("gazdagret").unwrap();
-    let osm_housenumbers = context::tests::TestFileSystem::make_file();
-    let html_cache = context::tests::TestFileSystem::make_file();
-    let mut file_system = context::tests::TestFileSystem::new();
-    let files = context::tests::TestFileSystem::make_files(
-        &ctx,
-        &[
-            (
-                "workdir/street-housenumbers-gazdagret.csv",
-                &osm_housenumbers,
-            ),
-            ("workdir/gazdagret.htmlcache.en", &html_cache),
-        ],
-    );
-    file_system.set_files(&files);
-    let mut mtimes: HashMap<String, Rc<RefCell<f64>>> = HashMap::new();
-    mtimes.insert(
-        ctx.get_abspath("workdir/street-housenumbers-gazdagret.csv"),
-        Rc::new(RefCell::new(1_f64)),
-    );
-    mtimes.insert(
-        ctx.get_abspath("workdir/gazdagret.htmlcache.en"),
-        Rc::new(RefCell::new(0_f64)),
-    );
-    file_system.set_mtimes(&mtimes);
-    let file_system_arc: Arc<dyn context::FileSystem> = Arc::new(file_system);
-    ctx.set_file_system(&file_system_arc);
-
-    let ret = is_missing_housenumbers_html_cached(&ctx, &relation).unwrap();
-
-    assert_eq!(ret, false);
-}
-
-/// Tests is_missing_housenumbers_html_cached(): the case when ref_housenumbers is new, so the cache entry is old.
-#[test]
-fn test_is_missing_housenumbers_html_cached_ref_housenumbers_new() {
-    let mut ctx = context::tests::make_test_context().unwrap();
-    let ref_housenumbers = context::tests::TestFileSystem::make_file();
-    let html_cache = context::tests::TestFileSystem::make_file();
-    let mut file_system = context::tests::TestFileSystem::new();
-    let files = context::tests::TestFileSystem::make_files(
-        &ctx,
-        &[
-            (
-                "workdir/street-housenumbers-reference-gazdagret.lst",
-                &ref_housenumbers,
-            ),
-            ("workdir/gazdagret.htmlcache.en", &html_cache),
-        ],
-    );
-    file_system.set_files(&files);
-    let mut mtimes: HashMap<String, Rc<RefCell<f64>>> = HashMap::new();
-    mtimes.insert(
-        ctx.get_abspath("workdir/street-housenumbers-reference-gazdagret.lst"),
-        Rc::new(RefCell::new(1_f64)),
-    );
-    mtimes.insert(
-        ctx.get_abspath("workdir/gazdagret.htmlcache.en"),
-        Rc::new(RefCell::new(0_f64)),
-    );
-    file_system.set_mtimes(&mtimes);
-    let file_system_arc: Arc<dyn context::FileSystem> = Arc::new(file_system);
-    ctx.set_file_system(&file_system_arc);
-    let mut relations = areas::Relations::new(&ctx).unwrap();
-    let relation = relations.get_relation("gazdagret").unwrap();
-
-    let ret = is_missing_housenumbers_html_cached(&ctx, &relation).unwrap();
-
-    assert_eq!(ret, false);
-}
-
-/// Tests is_missing_housenumbers_html_cached(): the case when relation is new, so the cache entry is old.
-#[test]
-fn test_is_missing_housenumbers_html_cached_relation_new() {
-    let mut ctx = context::tests::make_test_context().unwrap();
-    let osm_streets = context::tests::TestFileSystem::make_file();
-    let osm_housenumbers = context::tests::TestFileSystem::make_file();
-    let ref_housenumbers = context::tests::TestFileSystem::make_file();
-    let html_cache = context::tests::TestFileSystem::make_file();
-    let relation_file = context::tests::TestFileSystem::make_file();
-    let mut file_system = context::tests::TestFileSystem::new();
-    let files = context::tests::TestFileSystem::make_files(
-        &ctx,
-        &[
-            ("workdir/streets-gazdagret.csv", &osm_streets),
-            (
-                "workdir/street-housenumbers-gazdagret.csv",
-                &osm_housenumbers,
-            ),
-            (
-                "workdir/street-housenumbers-reference-gazdagret.lst",
-                &ref_housenumbers,
-            ),
-            ("workdir/gazdagret.htmlcache.en", &html_cache),
-            ("data/relation-gazdagret.yaml", &relation_file),
-        ],
-    );
-    file_system.set_files(&files);
-    file_system
-        .write_from_string("cached", &ctx.get_abspath("workdir/gazdagret.htmlcache.en"))
-        .unwrap();
-    let mut mtimes: HashMap<String, Rc<RefCell<f64>>> = HashMap::new();
-    mtimes.insert(
-        ctx.get_abspath("workdir/streets-gazdagret.csv"),
-        Rc::new(RefCell::new(0_f64)),
-    );
-    mtimes.insert(
-        ctx.get_abspath("workdir/street-housenumbers-gazdagret.csv"),
-        Rc::new(RefCell::new(0_f64)),
-    );
-    mtimes.insert(
-        ctx.get_abspath("workdir/street-housenumbers-reference-gazdagret.lst"),
-        Rc::new(RefCell::new(0_f64)),
-    );
-    mtimes.insert(
-        ctx.get_abspath("workdir/gazdagret.htmlcache.en"),
-        Rc::new(RefCell::new(0_f64)),
-    );
-    mtimes.insert(
-        ctx.get_abspath("data/relation-gazdagret.yaml"),
-        Rc::new(RefCell::new(1_f64)),
-    );
-    file_system.set_mtimes(&mtimes);
-    let file_system_arc: Arc<dyn context::FileSystem> = Arc::new(file_system);
-    ctx.set_file_system(&file_system_arc);
-    let mut relations = areas::Relations::new(&ctx).unwrap();
-    let relation = relations.get_relation("gazdagret").unwrap();
-
-    let ret = is_missing_housenumbers_html_cached(&ctx, &relation).unwrap();
-
-    assert_eq!(ret, false);
-}
-
 /// Tests get_additional_housenumbers_html(): the case when we find the result in cache
 #[test]
 fn test_get_additional_housenumbers_html() {
@@ -398,74 +190,12 @@ fn test_get_missing_housenumbers_json() {
     assert_eq!(ret, "{'cached':'yes'}");
 }
 
-/// Tests get_missing_housenumbers_html().
-#[test]
-fn test_get_missing_housenumbers_html() {
-    let mut ctx = context::tests::make_test_context().unwrap();
-    let osm_streets = context::tests::TestFileSystem::make_file();
-    let osm_housenumbers = context::tests::TestFileSystem::make_file();
-    let ref_housenumbers = context::tests::TestFileSystem::make_file();
-    let html_cache = context::tests::TestFileSystem::make_file();
-    let relation_file = context::tests::TestFileSystem::make_file();
-    let mut file_system = context::tests::TestFileSystem::new();
-    let files = context::tests::TestFileSystem::make_files(
-        &ctx,
-        &[
-            ("workdir/streets-gazdagret.csv", &osm_streets),
-            (
-                "workdir/street-housenumbers-gazdagret.csv",
-                &osm_housenumbers,
-            ),
-            (
-                "workdir/street-housenumbers-reference-gazdagret.lst",
-                &ref_housenumbers,
-            ),
-            ("workdir/gazdagret.htmlcache.en", &html_cache),
-            ("data/relation-gazdagret.yaml", &relation_file),
-        ],
-    );
-    file_system.set_files(&files);
-    file_system
-        .write_from_string("cached", &ctx.get_abspath("workdir/gazdagret.htmlcache.en"))
-        .unwrap();
-    let mut mtimes: HashMap<String, Rc<RefCell<f64>>> = HashMap::new();
-    mtimes.insert(
-        ctx.get_abspath("workdir/streets-gazdagret.csv"),
-        Rc::new(RefCell::new(0_f64)),
-    );
-    mtimes.insert(
-        ctx.get_abspath("workdir/street-housenumbers-gazdagret.csv"),
-        Rc::new(RefCell::new(0_f64)),
-    );
-    mtimes.insert(
-        ctx.get_abspath("workdir/street-housenumbers-reference-gazdagret.lst"),
-        Rc::new(RefCell::new(0_f64)),
-    );
-    mtimes.insert(
-        ctx.get_abspath("workdir/gazdagret.htmlcache.en"),
-        Rc::new(RefCell::new(1_f64)),
-    );
-    mtimes.insert(
-        ctx.get_abspath("data/relation-gazdagret.yaml"),
-        Rc::new(RefCell::new(0_f64)),
-    );
-    file_system.set_mtimes(&mtimes);
-    let file_system_arc: Arc<dyn context::FileSystem> = Arc::new(file_system);
-    ctx.set_file_system(&file_system_arc);
-    let mut relations = areas::Relations::new(&ctx).unwrap();
-    let mut relation = relations.get_relation("gazdagret").unwrap();
-
-    let ret = get_missing_housenumbers_html(&ctx, &mut relation).unwrap();
-
-    assert_eq!(ret.get_value(), "cached");
-}
-
 /// Tests is_cache_current()
 #[test]
 fn test_is_cache_current() {
     let mut ctx = context::tests::make_test_context().unwrap();
     let mut file_system = context::tests::TestFileSystem::new();
-    let cache_path = "workdir/gazdagret.htmlcache.en";
+    let cache_path = "workdir/gazdagret.json.cache";
     file_system.set_hide_paths(&[cache_path.to_string()]);
     let file_system_arc: Arc<dyn context::FileSystem> = Arc::new(file_system);
     ctx.set_file_system(&file_system_arc);
