+++ src/wsgi/tests.rs
@@ -383,23 +383,17 @@ fn test_missing_housenumbers_well_formed() {
     });
     let yamls_cache_value = context::tests::TestFileSystem::write_json_to_file(&yamls_cache);
     let percent_value = context::tests::TestFileSystem::make_file();
-    let html_cache_value = context::tests::TestFileSystem::make_file();
     let json_cache_value = context::tests::TestFileSystem::make_file();
     let files = context::tests::TestFileSystem::make_files(
         &test_wsgi.ctx,
         &[
             ("data/yamls.cache", &yamls_cache_value),
             ("workdir/gazdagret.percent", &percent_value),
-            ("workdir/gazdagret.htmlcache.en", &html_cache_value),
             ("workdir/gazdagret.cache.json", &json_cache_value),
         ],
     );
     file_system.set_files(&files);
     let mut mtimes: HashMap<String, Rc<RefCell<f64>>> = HashMap::new();
-    mtimes.insert(
-        test_wsgi.ctx.get_abspath("workdir/gazdagret.htmlcache.en"),
-        Rc::new(RefCell::new(0_f64)),
-    );
     mtimes.insert(
         test_wsgi.ctx.get_abspath("workdir/gazdagret.cache.json"),
         Rc::new(RefCell::new(0_f64)),
@@ -436,7 +430,6 @@ fn test_missing_housenumbers_compat() {
     let mut test_wsgi = TestWsgi::new();
     let mut file_system = context::tests::TestFileSystem::new();
     let streets_value = context::tests::TestFileSystem::make_file();
-    let htmlcache_value = context::tests::TestFileSystem::make_file();
     let jsoncache_value = context::tests::TestFileSystem::make_file();
     let yamls_cache = serde_json::json!({
         "relations.yaml": {
@@ -451,17 +444,12 @@ fn test_missing_housenumbers_compat() {
         &[
             ("data/yamls.cache", &yamls_cache_value),
             ("workdir/gazdagret.percent", &streets_value),
-            ("workdir/gazdagret.htmlcache.en", &htmlcache_value),
             ("workdir/gazdagret.cache.json", &jsoncache_value),
         ],
     );
     file_system.set_files(&files);
     // Make sure the cache is outdated.
     let mut mtimes: HashMap<String, Rc<RefCell<f64>>> = HashMap::new();
-    mtimes.insert(
-        test_wsgi.ctx.get_abspath("workdir/gazdagret.htmlcache.en"),
-        Rc::new(RefCell::new(0_f64)),
-    );
     mtimes.insert(
         test_wsgi.ctx.get_abspath("workdir/gazdagret.cache.json"),
         Rc::new(RefCell::new(0_f64)),
@@ -474,10 +462,6 @@ fn test_missing_housenumbers_compat() {
 
     {
         let mut guard = streets_value.borrow_mut();
-        assert_eq!(guard.seek(SeekFrom::Current(0)).unwrap() > 0, true);
-    }
-    {
-        let mut guard = htmlcache_value.borrow_mut();
         assert_eq!(guard.seek(SeekFrom::Current(0)).unwrap() > 0, true);
     }
     {
@@ -505,14 +489,12 @@ fn test_missing_housenumbers_compat_relation() {
         }
     });
     let yamls_cache_value = context::tests::TestFileSystem::write_json_to_file(&yamls_cache);
-    let htmlcache_value = context::tests::TestFileSystem::make_file();
     let jsoncache_value = context::tests::TestFileSystem::make_file();
     let percent_value = context::tests::TestFileSystem::make_file();
     let files = context::tests::TestFileSystem::make_files(
         &test_wsgi.ctx,
         &[
             ("data/yamls.cache", &yamls_cache_value),
-            ("workdir/budafok.htmlcache.en", &htmlcache_value),
             ("workdir/budafok.cache.json", &jsoncache_value),
             ("workdir/budafok.percent", &percent_value),
         ],
@@ -520,10 +502,6 @@ fn test_missing_housenumbers_compat_relation() {
     let mut file_system = context::tests::TestFileSystem::new();
     file_system.set_files(&files);
     let mut mtimes: HashMap<String, Rc<RefCell<f64>>> = HashMap::new();
-    mtimes.insert(
-        test_wsgi.ctx.get_abspath("workdir/budafok.htmlcache.en"),
-        Rc::new(RefCell::new(0_f64)),
-    );
     mtimes.insert(
         test_wsgi.ctx.get_abspath("workdir/budafok.cache.json"),
         Rc::new(RefCell::new(0_f64)),
